# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Time Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [time-entries]
      summary: List time entries
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: taskId
          in: query
          schema:
            type: integer
        - name: projectId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: billable
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [time-entries]
      summary: Log time
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{entryId}:
    get:
      operationId: getTimeEntry
      tags: [time-entries]
      summary: Get a time entry
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: Time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTimeEntry
      tags: [time-entries]
      summary: Update a time entry
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    delete:
      operationId: deleteTimeEntry
      tags: [time-entries]
      summary: Delete a time entry
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /time-entries/{entryId}/split:
    post:
      operationId: splitTimeEntry
      tags: [time-entries]
      summary: Split entry hours into billable and non-billable (Associate+)
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitHoursInput"
      responses:
        "200":
          description: Updated time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{entryId}/approve:
    post:
      operationId: approveTimeEntry
      tags: [time-entries]
      summary: Approve a time entry
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{entryId}/reject:
    post:
      operationId: rejectTimeEntry
      tags: [time-entries]
      summary: Reject a time entry
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{entryId}/reopen:
    post:
      operationId: reopenTimeEntry
      tags: [time-entries]
      summary: Reopen an approved time entry so it can be edited again
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: Reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"

  # ─── Dashboard ───────────────────────────────────────────────────────────
...
components:
  parameters:
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ────────────────────────────────────────────────────────────
...
    TimeEntry:
      type: object
      required: [id, userId, userName, taskId, taskName, hours, date, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
        userTitle:
          type: ["string", "null"]
          description: >-
            Overrides the rank label where the person's designation differs from it. Null shows the ordinary role label.
        taskId:
          type: integer
        taskName:
          type: string
        projectId:
          type: ["integer", "null"]
          description: "null for legacy entries logged before tasks became a global catalog"
        projectName:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        hours:
          type: number
        date:
          type: string
          format: date
        description:
          type: ["string", "null"]
        billableHours:
          type: ["number", "null"]
          description: "null = not yet split; 0..hours = explicitly split by Associate+"
        nonBillableHours:
          type: ["number", "null"]
          description: "computed: hours - billableHours; null if not yet split"
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedById:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TimeEntryInput:
      type: object
      required: [projectId, taskId, hours, date]
      properties:
        projectId:
          type: integer
        taskId:
          type: integer
        hours:
          type: number
          minimum: 0.25
          maximum: 24
        date:
          type: string
          format: date
        description:
          type: string
    TimeEntryUpdate:
      type: object
      properties:
        projectId:
          type: integer
        taskId:
          type: integer
        hours:
          type: number
          minimum: 0.25
          maximum: 24
        date:
          type: string
          format: date
        description:
          type: ["string", "null"]
    SplitHoursInput:
      type: object
      required: [billableHours]
      properties:
        billableHours:
          type: number
          minimum: 0
          description: "Hours to mark as billable; remainder becomes non-billable"

    # ─── Assignments ──────────────────────────────────────────────────────
...